ad generator/requirements.txt
groq==0.4.2
huggingface_hub==0.21.4
python-dotenv==1.0.1
